requirements.txt
flask==2.3.3
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0
pymupdf==1.23.8
werkzeug==2.3.7
tabula-py==2.7.0
pandas>=1.5.0
Pillow>=9.0.0 